tests/wpt/tests/css/css-toggle/parsing/toggle-visibility-valid.tentative.html

<!DOCTYPE HTML>
<meta charset="UTF-8">
<title>CSS Toggles: parsing toggle-visibility with valid values</title>
<link rel="author" title="L. David Baron" href="https://dbaron.org/">
<link rel="author" title="Google" href="http://www.google.com/">
<link rel="help" href="https://tabatkins.github.io/css-toggle/#toggle-visibility-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<body>
<script>

test_valid_value('toggle-visibility', 'normal');
test_valid_value('toggle-visibility', 'toggle t');
test_valid_value('toggle-visibility', 'toggle toggle');

</script>
